[EXPERIMENTAL] Add context to "callable" options
In order to support a more ergonomic interface while performing key
rotations, this passes additional contextual information to callable
values for options of encrypted attributes.

Here we pass the name of the attribute being encrypted as well as its
encryption options to the callable object.

For example:

```ruby
attr_encrypted(
  :ssn,
  key:  proc do |record, attribute, config|
    # record:    the record on which the `#ssn` attribute is being
    #            en/decrypted
    #
    # attribute: the name of the attribute being en/decrypted. In this
    #            case, :ssn
    #
    # config:    the hash from the model's encrypted_attributes map for
    #            :ssn.
  end,
)
```